수정 - 피드, 댓글 삭제 시 유저 유저 검증단계 추가 / front에서 유저정보가 일치할때만 api 버튼이 보이지만 url로 요청하면 지워지는 이슈가 있어서 서버에서 검증 #21 #28

+++ feed/services/comment_service.py
@@ -11,9 +11,9 @@ def get_comment_list(feed_id: int, offset: int, limit: int) -> QuerySet[FeedComm
     return FeedComment.objects.filter(feed_id=feed_id).order_by("id")[offset : offset + limit]
 
 
-def delete_an_comment(comment_id: int) -> None:
-    FeedComment.objects.filter(id=comment_id).delete()
+def delete_an_comment(user_id: int, comment_id: int) -> None:
+    FeedComment.objects.filter(id=comment_id, user_id=user_id).delete()
 
 
-def update_an_comment(comment_id: int, content: str) -> int:
-    return FeedComment.objects.filter(id=comment_id).update(content=content)
+def update_an_comment(user_id: int, comment_id: int, content: str) -> int:
+    return FeedComment.objects.filter(id=comment_id, user_id=user_id).update(content=content)
